src/requirements.txt
streamlit
pytesseract 
pillow 
langchain-google-genai 
langchain-core 
streamlit-back-camera-input
fpdf2
edge-tts